refactor(igdb): split god class into layered files
770-line IgdbApi becomes a facade over four focused sub-APIs (http client
with Twitch OAuth, games, platforms, genres) plus a shared types file.
Public surface — constructor, provider, methods, exceptions, callback,
maxMultiQueryBatch — is preserved 1:1, so 18 call sites and 9 test files
work without changes. Mirrors the earlier AniList split.
