+++ lib/core/api/igdb_api.dart
@@ -1,11 +1,16 @@
 import 'package:dio/dio.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:logging/logging.dart';
 
-import 'api_error_detail.dart';
 import '../services/api_key_initializer.dart';
 import '../../shared/models/game.dart';
 import '../../shared/models/platform.dart';
+import 'igdb/igdb_games_api.dart';
+import 'igdb/igdb_genres_api.dart';
+import 'igdb/igdb_http_client.dart';
+import 'igdb/igdb_platforms_api.dart';
+import 'igdb/igdb_types.dart';
+
+export 'igdb/igdb_types.dart';
 
 // Credentials seeded from apiKeysProvider loaded in main() before runApp().
 final Provider<IgdbApi> igdbApiProvider = Provider<IgdbApi>((Ref ref) {
@@ -21,224 +26,63 @@ final Provider<IgdbApi> igdbApiProvider = Provider<IgdbApi>((Ref ref) {
   return api;
 });
 
-class TwitchAuthResult {
-  const TwitchAuthResult({
-    required this.accessToken,
-    required this.expiresIn,
-    required this.tokenType,
-  });
-
-  factory TwitchAuthResult.fromJson(Map<String, dynamic> json) {
-    return TwitchAuthResult(
-      accessToken: json['access_token'] as String,
-      expiresIn: json['expires_in'] as int,
-      tokenType: json['token_type'] as String,
-    );
-  }
-
-  final String accessToken;
-  final int expiresIn;
-  final String tokenType;
-
-  /// Unix timestamp when the token expires.
-  int get expiresAt {
-    return DateTime.now().millisecondsSinceEpoch ~/ 1000 + expiresIn;
-  }
-}
-
-class IgdbApiException implements Exception {
-  const IgdbApiException(this.message, {this.statusCode, this.detail});
-
-  final String message;
-  final int? statusCode;
-  final String? detail;
-
-  @override
-  String toString() => 'IgdbApiException: $message (status: $statusCode)';
-}
-
-// Twitch OAuth + IGDB v4. Docs: https://api-docs.igdb.com/
-typedef IgdbTokenRefreshedCallback = void Function(
-  String accessToken,
-  int expiresAt,
-);
-
+/// IGDB v4 facade. See `igdb/` for layer breakdown:
+/// `igdb_http_client` (transport+auth), `igdb_games_api`, `igdb_platforms_api`,
+/// `igdb_genres_api`, `igdb_types` (DTOs).
 class IgdbApi {
-  IgdbApi({Dio? dio})
-      : _dio = dio ??
-            Dio(BaseOptions(
-              connectTimeout: _timeout,
-              receiveTimeout: _timeout,
-            ));
+  IgdbApi({Dio? dio}) : _client = IgdbHttpClient(dio: dio) {
+    _games = IgdbGamesApi(_client);
+    _platforms = IgdbPlatformsApi(_client);
+    _genres = IgdbGenresApi(_client);
+  }
 
-  static final Logger _log = Logger('IgdbApi');
+  final IgdbHttpClient _client;
+  late final IgdbGamesApi _games;
+  late final IgdbPlatformsApi _platforms;
+  late final IgdbGenresApi _genres;
 
-  static const Duration _timeout = Duration(seconds: 5);
-  static const String _twitchAuthUrl = 'https://id.twitch.tv/oauth2/token';
-  static const String _igdbBaseUrl = 'https://api.igdb.com/v4';
+  static const int maxMultiQueryBatch = IgdbGamesApi.maxMultiQueryBatch;
 
-  final Dio _dio;
-
-  String? _clientId;
-  String? _clientSecret;
-  String? _accessToken;
-
-  /// Invoked on auto-refresh so callers can persist the new token.
-  IgdbTokenRefreshedCallback? onTokenRefreshed;
-
-  bool _isRefreshing = false;
+  IgdbTokenRefreshedCallback? get onTokenRefreshed => _client.onTokenRefreshed;
+  set onTokenRefreshed(IgdbTokenRefreshedCallback? cb) {
+    _client.onTokenRefreshed = cb;
+  }
 
   void setCredentials({
     required String clientId,
     required String accessToken,
     String? clientSecret,
-  }) {
-    _clientId = clientId;
-    _clientSecret = clientSecret;
-    _accessToken = accessToken;
-  }
+  }) =>
+      _client.setCredentials(
+        clientId: clientId,
+        accessToken: accessToken,
+        clientSecret: clientSecret,
+      );
 
-  void clearCredentials() {
-    _clientId = null;
-    _clientSecret = null;
-    _accessToken = null;
-  }
+  void clearCredentials() => _client.clearCredentials();
 
   Future<TwitchAuthResult> getAccessToken({
     required String clientId,
     required String clientSecret,
-  }) async {
-    try {
-      final Response<dynamic> response = await _dio.post<dynamic>(
-        _twitchAuthUrl,
-        queryParameters: <String, dynamic>{
-          'client_id': clientId,
-          'client_secret': clientSecret,
-          'grant_type': 'client_credentials',
-        },
+  }) =>
+      _client.getAccessToken(
+        clientId: clientId,
+        clientSecret: clientSecret,
       );
-
-      if (response.statusCode == 200 && response.data != null) {
-        return TwitchAuthResult.fromJson(
-          response.data as Map<String, dynamic>,
-        );
-      }
-
-      throw IgdbApiException(
-        'Failed to get access token',
-        statusCode: response.statusCode,
-      );
-    } on DioException catch (e) {
-      final int? statusCode = e.response?.statusCode;
-      String message = 'Authentication failed';
-
-      if (statusCode == 400 || statusCode == 401) {
-        message = 'Invalid client ID or client secret';
-      } else if (e.type == DioExceptionType.connectionTimeout ||
-          e.type == DioExceptionType.receiveTimeout) {
-        message = 'Connection timeout';
-      } else if (e.type == DioExceptionType.connectionError) {
-        message = 'No internet connection';
-      }
-
-      throw IgdbApiException(
-        message,
-        statusCode: statusCode,
-        detail: buildApiErrorDetail(
-          apiName: 'IGDB Auth',
-          exception: e,
-          userMessage: message,
-        ),
-      );
-    }
-  }
 
   Future<bool> validateCredentials({
     required String clientId,
     required String clientSecret,
-  }) async {
-    try {
-      await getAccessToken(clientId: clientId, clientSecret: clientSecret);
-      return true;
-    } on IgdbApiException {
-      return false;
-    }
-  }
-
-  Future<List<Platform>> fetchPlatforms() async {
-    _ensureCredentials();
-
-    try {
-      final List<Platform> allPlatforms = <Platform>[];
-      int offset = 0;
-      const int limit = 500;
-
-      while (true) {
-        final Response<dynamic> response = await _igdbPost(
-          '/platforms',
-          data: 'fields id,name,abbreviation; limit $limit; offset $offset;',
-        );
-
-        if (response.statusCode != 200 || response.data == null) {
-          throw IgdbApiException(
-            'Failed to fetch platforms',
-            statusCode: response.statusCode,
-          );
-        }
-
-        final List<dynamic> data = response.data as List<dynamic>;
-        if (data.isEmpty) break;
-
-        final List<Platform> platforms = data
-            .map((dynamic item) =>
-                Platform.fromJson(item as Map<String, dynamic>))
-            .toList();
-
-        allPlatforms.addAll(platforms);
-
-        if (data.length < limit) break;
-        offset += limit;
-      }
-
-      return allPlatforms;
-    } on DioException catch (e) {
-      throw _handleDioException(e, 'Failed to fetch platforms');
-    }
-  }
-
-  Future<List<Platform>> fetchPlatformsByIds(List<int> ids) async {
-    if (ids.isEmpty) return <Platform>[];
-    _ensureCredentials();
-
-    try {
-      final String idList = ids.join(',');
-      final Response<dynamic> response = await _igdbPost(
-        '/platforms',
-        data:
-            'fields id,name,abbreviation; where id = ($idList); limit 500;',
+  }) =>
+      _client.validateCredentials(
+        clientId: clientId,
+        clientSecret: clientSecret,
       );
 
-      if (response.statusCode != 200 || response.data == null) {
-        throw IgdbApiException(
-          'Failed to fetch platforms by IDs',
-          statusCode: response.statusCode,
-        );
-      }
+  Future<List<Platform>> fetchPlatforms() => _platforms.fetchPlatforms();
 
-      final List<dynamic> data = response.data as List<dynamic>;
-      return data
-          .map((dynamic item) =>
-              Platform.fromJson(item as Map<String, dynamic>))
-          .toList();
-    } on DioException catch (e) {
-      throw _handleDioException(e, 'Failed to fetch platforms by IDs');
-    }
-  }
-
-  static const String _gameFields = '''
-    fields id, name, summary, rating, rating_count, first_release_date,
-           cover.image_id, artworks.image_id, genres.name, platforms, url;
-  ''';
+  Future<List<Platform>> fetchPlatformsByIds(List<int> ids) =>
+      _platforms.fetchPlatformsByIds(ids);
 
   Future<List<Game>> searchGames({
     required String query,
@@ -250,352 +94,42 @@ class IgdbApi {
     (int, int)? decade,
     int limit = 20,
     int offset = 0,
-  }) async {
-    _ensureCredentials();
-
-    if (query.trim().isEmpty) {
-      return <Game>[];
-    }
-
-    try {
-      final String escapedQuery = query.replaceAll('"', '\\"');
-
-      // IGDB query order: fields -> where -> search -> limit.
-      final StringBuffer body = StringBuffer(_gameFields);
-
-      // IGDB: `field = (a,b)` is ANY-of (OR match).
-      final List<String> conditions = <String>[];
-      if (platformIds != null && platformIds.isNotEmpty) {
-        conditions.add('platforms = (${platformIds.join(",")})');
-      }
-      if (genreIds != null && genreIds.isNotEmpty) {
-        conditions.add('genres = (${genreIds.join(",")})');
-      }
-      if (gameModeIds != null && gameModeIds.isNotEmpty) {
-        conditions.add('game_modes = (${gameModeIds.join(",")})');
-      }
-      if (minRating != null) {
-        conditions.add('rating >= $minRating');
-      }
-      if (year != null) {
-        final int start =
-            DateTime(year).millisecondsSinceEpoch ~/ 1000;
-        final int end =
-            DateTime(year + 1).millisecondsSinceEpoch ~/ 1000;
-        conditions.add(
-          'first_release_date >= $start & first_release_date < $end',
-        );
-      } else if (decade != null) {
-        final int start =
-            DateTime(decade.$1).millisecondsSinceEpoch ~/ 1000;
-        final int end =
-            DateTime(decade.$2 + 1).millisecondsSinceEpoch ~/ 1000;
-        conditions.add(
-          'first_release_date >= $start & first_release_date < $end',
-        );
-      }
-      if (conditions.isNotEmpty) {
-        body.write(' where ${conditions.join(" & ")};');
-      }
-
-      body.write(' search "$escapedQuery"; limit $limit;');
-      if (offset > 0) {
-        body.write(' offset $offset;');
-      }
-
-      final Response<dynamic> response = await _igdbPost(
-        '/games',
-        data: body.toString(),
+  }) =>
+      _games.searchGames(
+        query: query,
+        genreIds: genreIds,
+        platformIds: platformIds,
+        gameModeIds: gameModeIds,
+        minRating: minRating,
+        year: year,
+        decade: decade,
+        limit: limit,
+        offset: offset,
       );
-
-      if (response.statusCode != 200 || response.data == null) {
-        throw IgdbApiException(
-          'Failed to search games',
-          statusCode: response.statusCode,
-        );
-      }
-
-      final List<dynamic> data = response.data as List<dynamic>;
-      return data
-          .map((dynamic item) => Game.fromJson(item as Map<String, dynamic>))
-          .toList();
-    } on DioException catch (e) {
-      throw _handleDioException(e, 'Failed to search games');
-    }
-  }
-
-  /// IGDB multiquery cap: 10 sub-queries per request.
-  static const int maxMultiQueryBatch = 10;
-
-  static const int _multiSearchLimit = 20;
 
   Future<Map<int, List<Game>>> multiSearchGamesByName(
     List<({String name, int? platformId})> queries,
-  ) async {
-    if (queries.isEmpty) return <int, List<Game>>{};
-    _ensureCredentials();
+  ) =>
+      _games.multiSearchGamesByName(queries);
 
-    const String fields =
-        'fields id,name,summary,rating,rating_count,first_release_date,'
-        'cover.image_id,genres.name,platforms,url;';
+  Future<Map<String, Game>> lookupSteamGames(List<String> steamAppIds) =>
+      _games.lookupSteamGames(steamAppIds);
 
-    try {
-      final StringBuffer body = StringBuffer();
-      for (int i = 0; i < queries.length; i++) {
-        final String escaped = queries[i]
-            .name
-            .replaceAll('"', '\\"')
-            .replaceAll('*', '');
-        final String platformFilter = queries[i].platformId != null
-            ? ' & platforms = (${queries[i].platformId})'
-            : '';
-        body.writeln(
-          'query games "q_$i" { $fields '
-          'where name ~ *"$escaped"*$platformFilter; '
-          'limit $_multiSearchLimit; };',
-        );
-      }
+  Future<Game?> getGameById(int gameId) => _games.getGameById(gameId);
 
-      final Response<dynamic> response = await _igdbPost(
-        '/multiquery',
-        data: body.toString(),
-      );
-
-      if (response.statusCode != 200 || response.data == null) {
-        throw IgdbApiException(
-          'Failed to multi-search games',
-          statusCode: response.statusCode,
-        );
-      }
-
-      final List<dynamic> results = response.data as List<dynamic>;
-      final Map<int, List<Game>> mapped = <int, List<Game>>{};
-
-      for (final dynamic entry in results) {
-        final Map<String, dynamic> item = entry as Map<String, dynamic>;
-        final String name = item['name'] as String;
-        final int? index = int.tryParse(name.replaceFirst('q_', ''));
-        if (index == null) continue;
-
-        final List<dynamic> resultList =
-            (item['result'] as List<dynamic>?) ?? <dynamic>[];
-        mapped[index] = resultList
-            .map((dynamic g) => Game.fromJson(g as Map<String, dynamic>))
-            .toList();
-      }
-
-      for (int i = 0; i < queries.length; i++) {
-        mapped.putIfAbsent(i, () => <Game>[]);
-      }
-
-      return mapped;
-    } on DioException catch (e) {
-      throw _handleDioException(e, 'Failed to multi-search games');
-    }
-  }
-
-  /// IGDB `external_game_source` value for Steam.
-  static const int _steamSource = 1;
-
-  Future<Map<String, Game>> lookupSteamGames(
-    List<String> steamAppIds,
-  ) async {
-    if (steamAppIds.isEmpty) return <String, Game>{};
-    _ensureCredentials();
-
-    try {
-      // Step 1: Steam appId -> IGDB game id via external_games.
-      final Map<String, int> uidToGameId = <String, int>{};
-
-      for (int offset = 0; offset < steamAppIds.length; offset += 500) {
-        final List<String> batch = steamAppIds.sublist(
-          offset,
-          offset + 500 > steamAppIds.length
-              ? steamAppIds.length
-              : offset + 500,
-        );
-        final String uidList = batch.map((String id) => '"$id"').join(',');
-
-        final Response<dynamic> response = await _igdbPost(
-          '/external_games',
-          data: 'fields game,uid; '
-              'where external_game_source = $_steamSource '
-              '& uid = ($uidList); '
-              'limit 500;',
-        );
-
-        if (response.statusCode != 200 || response.data == null) {
-          throw IgdbApiException(
-            'Failed to lookup Steam games',
-            statusCode: response.statusCode,
-          );
-        }
-
-        final List<dynamic> data = response.data as List<dynamic>;
-        for (final dynamic item in data) {
-          final Map<String, dynamic> map = item as Map<String, dynamic>;
-          final String uid = map['uid'] as String;
-          final int gameId = map['game'] as int;
-          uidToGameId[uid] = gameId;
-        }
-      }
-
-      if (uidToGameId.isEmpty) return <String, Game>{};
-
-      // Step 2: fetch full game data by IGDB id (deduped).
-      final List<Game> games =
-          await getGamesByIds(uidToGameId.values.toSet().toList());
-
-      final Map<int, Game> gamesById = <int, Game>{
-        for (final Game game in games) game.id: game,
-      };
-
-      final Map<String, Game> result = <String, Game>{};
-      for (final MapEntry<String, int> entry in uidToGameId.entries) {
-        final Game? game = gamesById[entry.value];
-        if (game != null) {
-          result[entry.key] = game;
-        }
-      }
-
-      return result;
-    } on DioException catch (e) {
-      throw _handleDioException(e, 'Failed to lookup Steam games');
-    }
-  }
-
-  Future<Game?> getGameById(int gameId) async {
-    _ensureCredentials();
-
-    try {
-      final Response<dynamic> response = await _igdbPost(
-        '/games',
-        data: '$_gameFields where id = $gameId;',
-      );
-
-      if (response.statusCode != 200 || response.data == null) {
-        throw IgdbApiException(
-          'Failed to fetch game',
-          statusCode: response.statusCode,
-        );
-      }
-
-      final List<dynamic> data = response.data as List<dynamic>;
-      if (data.isEmpty) return null;
-
-      return Game.fromJson(data.first as Map<String, dynamic>);
-    } on DioException catch (e) {
-      throw _handleDioException(e, 'Failed to fetch game');
-    }
-  }
-
-  Future<List<Game>> getGamesByIds(List<int> gameIds) async {
-    _ensureCredentials();
-
-    if (gameIds.isEmpty) {
-      return <Game>[];
-    }
-
-    try {
-      // IGDB caps a single request at 500 records.
-      final List<Game> allGames = <Game>[];
-
-      for (int i = 0; i < gameIds.length; i += 500) {
-        final List<int> batch = gameIds.sublist(
-          i,
-          i + 500 > gameIds.length ? gameIds.length : i + 500,
-        );
-
-        final String idsString = batch.join(',');
-
-        final Response<dynamic> response = await _igdbPost(
-          '/games',
-          data: '$_gameFields where id = ($idsString); limit 500;',
-        );
-
-        if (response.statusCode != 200 || response.data == null) {
-          throw IgdbApiException(
-            'Failed to fetch games',
-            statusCode: response.statusCode,
-          );
-        }
-
-        final List<dynamic> data = response.data as List<dynamic>;
-        final List<Game> games = data
-            .map((dynamic item) => Game.fromJson(item as Map<String, dynamic>))
-            .toList();
-
-        allGames.addAll(games);
-      }
-
-      return allGames;
-    } on DioException catch (e) {
-      throw _handleDioException(e, 'Failed to fetch games');
-    }
-  }
+  Future<List<Game>> getGamesByIds(List<int> gameIds) =>
+      _games.getGamesByIds(gameIds);
 
   Future<List<Game>> getTopGamesByPlatform({
     required int platformId,
     int minRatingCount = 20,
     int limit = 50,
-  }) async {
-    _ensureCredentials();
-
-    try {
-      final StringBuffer body = StringBuffer(_gameFields);
-      body.write(
-        ' where platforms = ($platformId)'
-        ' & rating_count >= $minRatingCount'
-        ' & rating != null;',
+  }) =>
+      _games.getTopGamesByPlatform(
+        platformId: platformId,
+        minRatingCount: minRatingCount,
+        limit: limit,
       );
-      body.write(' sort rating desc;');
-      body.write(' limit $limit;');
-
-      final Response<dynamic> response = await _igdbPost(
-        '/games',
-        data: body.toString(),
-      );
-
-      if (response.statusCode != 200 || response.data == null) {
-        throw IgdbApiException(
-          'Failed to fetch top games',
-          statusCode: response.statusCode,
-        );
-      }
-
-      final List<dynamic> data = response.data as List<dynamic>;
-      return data
-          .map((dynamic item) => Game.fromJson(item as Map<String, dynamic>))
-          .toList();
-    } on DioException catch (e) {
-      throw _handleDioException(e, 'Failed to fetch top games');
-    }
-  }
-
-  Future<List<Map<String, dynamic>>> fetchGenres() async {
-    _ensureCredentials();
-
-    try {
-      final Response<dynamic> response = await _igdbPost(
-        '/genres',
-        data: 'fields id,name; limit 50; sort name asc;',
-      );
-
-      if (response.statusCode != 200 || response.data == null) {
-        throw IgdbApiException(
-          'Failed to fetch genres',
-          statusCode: response.statusCode,
-        );
-      }
-
-      final List<dynamic> data = response.data as List<dynamic>;
-      return data
-          .map((dynamic item) => item as Map<String, dynamic>)
-          .toList();
-    } on DioException catch (e) {
-      throw _handleDioException(e, 'Failed to fetch genres');
-    }
-  }
 
   Future<List<Game>> browseGames({
     List<int>? genreIds,
@@ -608,163 +142,21 @@ class IgdbApi {
     int limit = 20,
     int offset = 0,
     int minRatingCount = 10,
-  }) async {
-    _ensureCredentials();
-
-    try {
-      final StringBuffer where =
-          StringBuffer('where rating_count > $minRatingCount');
-
-      if (genreIds != null && genreIds.isNotEmpty) {
-        where.write(' & genres = (${genreIds.join(",")})');
-      }
-      if (platformIds != null && platformIds.isNotEmpty) {
-        where.write(' & platforms = (${platformIds.join(",")})');
-      }
-      if (gameModeIds != null && gameModeIds.isNotEmpty) {
-        where.write(' & game_modes = (${gameModeIds.join(",")})');
-      }
-      if (minRating != null) {
-        where.write(' & rating >= $minRating');
-      }
-      if (year != null) {
-        final int start =
-            DateTime(year).millisecondsSinceEpoch ~/ 1000;
-        final int end =
-            DateTime(year + 1).millisecondsSinceEpoch ~/ 1000;
-        where.write(
-          ' & first_release_date >= $start & first_release_date < $end',
-        );
-      } else if (decade != null) {
-        final int start =
-            DateTime(decade.$1).millisecondsSinceEpoch ~/ 1000;
-        final int end =
-            DateTime(decade.$2 + 1).millisecondsSinceEpoch ~/ 1000;
-        where.write(
-          ' & first_release_date >= $start & first_release_date < $end',
-        );
-      }
-
-      final StringBuffer body = StringBuffer(_gameFields);
-      body.write(' $where;');
-      body.write(' sort $sortBy;');
-      body.write(' limit $limit;');
-      if (offset > 0) {
-        body.write(' offset $offset;');
-      }
-
-      final Response<dynamic> response = await _igdbPost(
-        '/games',
-        data: body.toString(),
+  }) =>
+      _games.browseGames(
+        genreIds: genreIds,
+        platformIds: platformIds,
+        gameModeIds: gameModeIds,
+        minRating: minRating,
+        year: year,
+        decade: decade,
+        sortBy: sortBy,
+        limit: limit,
+        offset: offset,
+        minRatingCount: minRatingCount,
       );
 
-      if (response.statusCode != 200 || response.data == null) {
-        throw IgdbApiException(
-          'Failed to browse games',
-          statusCode: response.statusCode,
-        );
-      }
+  Future<List<Map<String, dynamic>>> fetchGenres() => _genres.fetchGenres();
 
-      final List<dynamic> data = response.data as List<dynamic>;
-      return data
-          .map((dynamic item) => Game.fromJson(item as Map<String, dynamic>))
-          .toList();
-    } on DioException catch (e) {
-      throw _handleDioException(e, 'Failed to browse games');
-    }
-  }
-
-  // Maps Dio errors to user-facing messages; 401 = bad/expired token, 429 = rate limit.
-  IgdbApiException _handleDioException(DioException e, String defaultMessage) {
-    final int? statusCode = e.response?.statusCode;
-    String message = defaultMessage;
-
-    if (statusCode == 401) {
-      message = 'Invalid or expired access token';
-    } else if (statusCode == 429) {
-      message = 'Rate limit exceeded. Please try again later';
-    } else if (e.type == DioExceptionType.connectionTimeout ||
-        e.type == DioExceptionType.receiveTimeout) {
-      message = 'Connection timeout';
-    } else if (e.type == DioExceptionType.connectionError) {
-      message = 'No internet connection';
-    }
-
-    return IgdbApiException(
-      message,
-      statusCode: statusCode,
-      detail: buildApiErrorDetail(
-        apiName: 'IGDB',
-        exception: e,
-        userMessage: message,
-      ),
-    );
-  }
-
-  // On 401, refresh the Twitch token once and retry the request.
-  Future<Response<dynamic>> _igdbPost(
-    String endpoint, {
-    required String data,
-  }) async {
-    try {
-      return await _dio.post<dynamic>(
-        '$_igdbBaseUrl$endpoint',
-        options: Options(
-          headers: <String, dynamic>{
-            'Client-ID': _clientId,
-            'Authorization': 'Bearer $_accessToken',
-          },
-        ),
-        data: data,
-      );
-    } on DioException catch (e) {
-      if (e.response?.statusCode == 401 && await _tryRefreshToken()) {
-        return _dio.post<dynamic>(
-          '$_igdbBaseUrl$endpoint',
-          options: Options(
-            headers: <String, dynamic>{
-              'Client-ID': _clientId,
-              'Authorization': 'Bearer $_accessToken',
-            },
-          ),
-          data: data,
-        );
-      }
-      rethrow;
-    }
-  }
-
-  // Guarded by _isRefreshing to avoid concurrent refresh storms on parallel 401s.
-  Future<bool> _tryRefreshToken() async {
-    if (_isRefreshing) return false;
-    if (_clientId == null || _clientSecret == null) return false;
-
-    _isRefreshing = true;
-    try {
-      _log.info('Auto-refreshing IGDB token...');
-      final TwitchAuthResult result = await getAccessToken(
-        clientId: _clientId!,
-        clientSecret: _clientSecret!,
-      );
-      _accessToken = result.accessToken;
-      onTokenRefreshed?.call(result.accessToken, result.expiresAt);
-      _log.info('IGDB token refreshed successfully');
-      return true;
-    } on IgdbApiException catch (e) {
-      _log.warning('IGDB token refresh failed: $e');
-      return false;
-    } finally {
-      _isRefreshing = false;
-    }
-  }
-
-  void _ensureCredentials() {
-    if (_clientId == null || _accessToken == null) {
-      throw const IgdbApiException('API credentials not set');
-    }
-  }
-
-  void dispose() {
-    _dio.close();
-  }
+  void dispose() => _client.dispose();
 }
